fix: remove typings file and add typings directly to components
