+++ projects/form-elements/src/exports.ts
@@ -3,13 +3,3 @@
  */
 export * from './lib/form-elements.module'
 export { ModuleOptions, OPTIONS } from './lib/options'
-export {
-    FValidationConfig,
-    FExplanationConfig,
-    FRowLayouts,
-    FColumnAlignments,
-    FSelectOption,
-    FSelectPlaceholder,
-    FRadiobuttonOption,
-    FButtonType,
-} from './lib/typings'
