+++ projects/form-elements/src/lib/f-row/f-row.component.ts
@@ -1,7 +1,9 @@
 import { Component, ViewEncapsulation, Input } from '@angular/core'
 import { modifiers } from '@pascaliske/html-helpers'
-import { FRowLayouts } from '../typings'
 
+/**
+ * F-Row
+ */
 @Component({
     selector: 'cmp-f-row',
     templateUrl: './f-row.component.html',
@@ -11,7 +13,7 @@ export class FRowComponent {
     public static readonly cmpName: string = 'FRowComponent'
 
     @Input()
-    public layout: FRowLayouts = '1'
+    public layout: '1' | '1-1' | '1-1-1' | '1-2' | '2-1' = '1'
 
     public constructor() {}
